refactor cost_estimator to cost and move inside benchmarking module
