+++ benchmarking/__init__.py
@@ -27,6 +27,7 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+from benchmarking import cost
 from benchmarking import job
 from benchmarking import manager
 from benchmarking import utils
